refactor: add orderOptions struct
