+++ x/ecocredit/server/marketplace/buy_direct.go
@@ -72,8 +72,12 @@ func (k Keeper) BuyDirect(ctx context.Context, req *marketplace.MsgBuyDirect) (*
 	}
 
 	// fill the order, updating balances and the sell order in state
-	if err = k.fillOrder(ctx, sellOrder, buyerAcc, creditOrderQty, coinCost, false, !req.DisableAutoRetire,
-		req.RetirementLocation, batch.BatchDenom); err != nil {
+	if err = k.fillOrder(ctx, sellOrder, buyerAcc, creditOrderQty, coinCost, orderOptions{
+		autoRetire:         !req.DisableAutoRetire,
+		canPartialFill:     false,
+		batchDenom:         batch.BatchDenom,
+		retirementLocation: req.RetirementLocation,
+	}); err != nil {
 		return nil, fmt.Errorf("error filling order: %w", err)
 	}
